Invoice fix
